refactor(ui): split ProcessTree D3 rendering into ProcessTree.render.ts
Consolidation pass (deferred candidate from #620). ProcessTree.tsx was the top cognitive-complexity file (~253), though with no S3776 gate violation: the complexity is spread across many sub-threshold functions and the file simply concentrated the D3 render logic beside the React state. Relocate that logic into a new sibling module so the component file carries only the component.

Moved into ProcessTree.render.ts (behavior byte-identical, a pure relocation): renderTree (exported), the 11 pure per-node helpers (computeLayoutBounds, evidenceNodeClass, nodeDot*, nodeHasChevron, chevronGlyph, nodeLabel*), the render-only constants, and the RenderResult / TreeInteractions interfaces (exported). ProcessTree.tsx goes 910 -> 647 lines and imports renderTree (+ TREE_MARGIN_PX, still used by zoomToNode) from the new module; now-unused imports (toD3Hierarchy, buildNodeTooltip, nodeEvidenceMarked) dropped. The dependency is one-way (render.ts never imports the .tsx).

Also fixes the one open Sonar issue on the file (S7735 "unexpected negated condition") in the moved chevronGlyph, flipped to positive-first with identical behavior.

No test imports needed re-pointing (only ProcessTreeView is imported by the suite). Verified: tsc, eslint, vitest (62 tests unchanged), and render.ts coverage 99% stmts / 100% lines via the existing component suite.
