+++ docs/maintenance/log.md
@@ -55,4 +55,5 @@ or change the cadence - running tasks with no record is the same as not running 
 2026-07-05  dead-code-sweep  done  no findings  Go: `staticcheck -checks=U1000` (upgraded to 2026.1/v0.7.0 for Go 1.26; the old v0.4.7 segfaulted) reports 0 unused first-party symbols across server/agent/internal (also CI-gated by golangci `unused`+`staticcheck`). TS: `ts-prune` 0 genuine dead exports (all 12 hits are "used in module" false positives). Config knobs already covered clean by config-surface-review. Routes not re-swept (issue-filing territory, unchanged since 05-19)
 2026-07-05  test-suite-health  done  no code changes  14 Go skips all legit env/mode/platform guards (MySQL/ClickHouse DSN, git, macOS/Windows, -short, PBT known-state, PATH), 0 TS/Swift skips, no stale perma-skips; `-race` single pass green (2367 tests, 0 fails, 0 data races); slowest ~14s (catalog ESF-fixture replays) + ~10-11s real-MySQL integration (SSO/trace-sampler/appconfig), essential not accidental (no sleeps; DSN-gated so unit runs stay fast); appcontrol list tests substantively cover the consolidation refactor, no tautology smells. Bounded to count=1 this cycle (last count=3 clean 2026-06-18, minimal churn since); no new thinness issues to file
 2026-07-05  consolidation-pass  done  branch chore/maintenance-consolidation-2026-07-05  baseline dup 0.6% (23 blocks / 618 lines), aggregate cog 6706 over 65712 ncloc (down from 0.7%/6751 pre-#618). Two clusters, both behavior byte-identical (existing PBT/integration suites assert each; [no-behavior-change]/[no-docs-change]). (1) Shared HTTP request/response primitives into `httpserver`: `ReadCappedBody`/`DecodeCappedJSON` (the limit+1 read + `len>limit` check, the exact invariant CodeRabbit flagged on #618) replace 5 hand-rolled copies across two contexts (operator appcontrol/detectionconfig + identity useradmin/saadmin/ssoadmin); `WriteJSONError` centralizes the identity `{"error":code}` envelope behind the 3 identical `writeErr` wrappers. (2) Cross-context token dup: the `v1.<payload>.<mac>` HMAC-SHA256 envelope (split/mac/Seal + the strict-decode + constant-time verify core) was mirrored in `endpoint/internal/signedtoken` and `identity/internal/satoken`; extracted to new `internal/hmactoken` (shared via internal/ per ADR-0004, not sideways), each package keeping its own Claims + post-verify checks (kid/audience/expiry) and mapping the shared sentinels back to its own. hmactoken 97.5% cov, signedtoken 92.6%, satoken 87.5%. Deferred to next passes: `ProcessTree.tsx` cog-253 (top hotspot, single-session-plus + just touched in #619, file as issue per task rule), operator two-field error-writer shape
+2026-07-05  consolidation-pass  done  branch chore/consolidation-processtree-operator-2026-07-05  cleared the two candidates deferred from the prior entry, both behavior byte-identical ([no-docs-change]). (1) `ProcessTree.tsx` was the top cognitive-complexity file (~253) but carried NO S3776 gate violation (the metric is spread across many sub-threshold functions; only OPEN issue was a MINOR S7735 negated-condition). Relocated the D3 render logic (`renderTree` + 11 pure per-node helpers + the render-only constants + `RenderResult`/`TreeInteractions`) into a new sibling `ProcessTree.render.ts` (910->647 lines; render.ts 271 lines, 99% stmt / 100% line cov via the existing component suite, 62 tests unchanged), lowering the .tsx file metric by moving the concern out; fixed the S7735 nit in the moved `chevronGlyph`. (2) operator two-field error envelope centralized in `writeOperatorErr` (handler.go); `writeAppControlErr`/`writeDetectionConfigErr` now delegate (all 56 call sites unchanged). Next hotspots unchanged.
